extract build l2
